+++ src/Middleware/NotifierMiddleware.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace SchedulerBundle\Middleware;
 
+use SchedulerBundle\SchedulerInterface;
 use SchedulerBundle\Task\TaskInterface;
 use SchedulerBundle\TaskBag\NotificationTaskBag;
 use Symfony\Component\Notifier\Notification\Notification;
@@ -13,13 +14,39 @@ use Symfony\Component\Notifier\Recipient\Recipient;
 /**
  * @author Guillaume Loulier <contact@guillaumeloulier.fr>
  */
-final class NotifierMiddleware implements PreExecutionMiddlewareInterface, PostExecutionMiddlewareInterface
+final class NotifierMiddleware implements PreSchedulingMiddlewareInterface, PostSchedulingMiddlewareInterface, PreExecutionMiddlewareInterface, PostExecutionMiddlewareInterface, OrderedMiddlewareInterface
 {
     private ?NotifierInterface $notifier;
 
     public function __construct(?NotifierInterface $notifier = null)
     {
         $this->notifier = $notifier;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function preScheduling(TaskInterface $task, SchedulerInterface $scheduler): void
+    {
+        if (!$task->getBeforeSchedulingNotificationBag() instanceof NotificationTaskBag) {
+            return;
+        }
+
+        $bag = $task->getBeforeSchedulingNotificationBag();
+        $this->notify($bag->getNotification(), $bag->getRecipients());
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function postScheduling(TaskInterface $task, SchedulerInterface $scheduler): void
+    {
+        if (!$task->getAfterSchedulingNotificationBag() instanceof NotificationTaskBag) {
+            return;
+        }
+
+        $bag = $task->getAfterSchedulingNotificationBag();
+        $this->notify($bag->getNotification(), $bag->getRecipients());
     }
 
     public function preExecute(TaskInterface $task): void
@@ -40,6 +67,14 @@ final class NotifierMiddleware implements PreExecutionMiddlewareInterface, PostE
 
         $bag = $task->getAfterExecutingNotificationBag();
         $this->notify($bag->getNotification(), $bag->getRecipients());
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getPriority(): int
+    {
+        return 2;
     }
 
     /**