+++ src/Task/TaskList.php
@@ -10,7 +10,6 @@ use InvalidArgumentException;
 use Throwable;
 use function array_filter;
 use function array_key_exists;
-use function array_map;
 use function array_values;
 use function count;
 use function gettype;