+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
 * Doctrine dependencies updated (see https://github.com/Guikingone/SchedulerBundle/pull/17)
+* Middleware introduced (https://github.com/Guikingone/SchedulerBundle/pull/19)
 
 0.2.0
 -----
@@ -12,7 +13,7 @@ CHANGELOG
 * Task notifications added (see https://github.com/Guikingone/SchedulerBundle/pull/1)
 * Task & Worker lifecycle logs added (see https://github.com/Guikingone/SchedulerBundle/pull/5)
 * PHP 7.2 & 7.3 support dropped (see https://github.com/Guikingone/SchedulerBundle/pull/13)
-* PHP 8.0 support added (see https://github.com/Guikingone/SschedulerBundle/pull/13)
+* PHP 8.0 support added (see https://github.com/Guikingone/SchedulerBundle/pull/13)
 
 0.1.0
 -----